Code to gradually converge on alarm time
Set up the code to check the bus time regularly as the alarm time approaches,
so that it can be triggered at the right moment.

So far the triggered alarm only appears in the log.

Unfortunately, this required a substantial change to the API access; we had to
implement a whole second version of the HTTP code, because the Ion library
simply never gave a response.  I suspect that this is because it doesn't work
when called from a non-GUI thread.