Manually trust OCTranspo intermediate CA
The OCTranspo API has changed to a cert signed by an intermediate CA,
but it does not offer that CA cert, nor is it in the Android root store,
thus the connection to the API was no longer trusted.

Using the new API 24 feature for network configuration, we can
explicitly include and trust that intermediate CA to make this work
again.

So far, this works only on Android 7.0+ (which is essentially no real
devices at this time).